requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
aniso8601==1.1.0
APScheduler==3.0.5
babelfish==0.5.5
beautifulsoup4==4.4.1
CherryPy==5.1.0
dj-database-url==0.4.0
Django==1.8
djangorestframework==3.3.2
feedparser==5.2.1
Flask==0.10.1
Flask-Compress==1.3.0
Flask-Cors==2.1.2
Flask-Login==0.3.2
Flask-RESTful==0.3.5
flask-restplus==0.8.6
FlexGet==1.2.479
functools32==3.2.3.post2
futures==3.0.5
guessit==2.0.4
gunicorn==19.4.5
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
MarkupSafe==0.23
ordereddict==1.1
path.py==8.1.2
Pillow==3.1.1
plumbum==1.6.1.post0
progressbar==2.3
pynzb==0.1.0
pyparsing==2.1.0
PyRSS2Gen==1.1
python-dateutil==2.5.0
pytvmaze==1.4.7
pytz==2015.7
PyYAML==3.11
rebulk==0.7.1
requests==2.9.1
rpyc==3.3.0
Safe==0.4
six==1.10.0
SQLAlchemy==1.0.12
tzlocal==1.2.2
Werkzeug==0.11.4
whitenoise==2.0.6
psycopg2==2.5.4
